requirements.txt
streamlit==1.31.1

langchain==0.1.16
langchain-core==0.1.52
langchain-community==0.0.38
langchain-groq==0.1.4
langchain-huggingface==0.0.3
langchain-text-splitters==0.0.1

chromadb==0.4.24
sentence-transformers==2.6.1
huggingface-hub >= 0.23.0

pypdf==4.0.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
webdriver-manager==4.0.1

pysqlite3-binary==0.5.2
tiktoken==0.5.2
posthog==3.0.2